Add support for Pool Investments - Splitting balances (#68)
* use a split balance to track cash/investments

* remove FixedPoint contract -> library note, now that it is done

* use a lib for Balance to calculate total()

* use SafeCast everywhere

* use single 128 bit ONE function to be upcast when dealing with uint256s

* pack bits in TokenData by using int128 for delta

* fix linter errors

* maintain vault balance as cash/invested

+++ contracts/IVault.sol
@@ -141,7 +141,7 @@ interface IVault {
     // Diffs array. Instead, the caller provides the indices for the Diffs array, leading to
     // gas savings.
     struct TokenData {
-        int256 delta;
-        uint256 tokenDiffIndex;
+        int128 delta;
+        uint128 tokenDiffIndex;
     }
 }
